+++ lib/index.ts
@@ -2,3 +2,5 @@ export * from './components';
 export * from './hooks';
 export * from './icons';
 export * from './interfaces';
+export * from './services';
+export * from './types';