+++ lib/services/index.ts
@@ -1,0 +1,4 @@
+export * from './clipboard.service';
+export * from './file.service';
+export * from './local.service';
+export * from './logger.service';