+++ src/main/examples/ServiceLoggerPage.tsx
@@ -1,0 +1,7 @@
+import { withOptions } from './components';
+
+const ServiceLoggerPageBase = (): JSX.Element => {
+	return <>Switch to "Usage" to see how to use this service</>;
+};
+
+export const ServiceLoggerPage: () => JSX.Element = withOptions(ServiceLoggerPageBase, {}, 'ServiceLoggerPageBase');