+++ src/main/examples/index.ts
@@ -47,4 +47,9 @@ export * from './HookUseOnDestroyPage';
 export * from './HookUsePreviousPage';
 export * from './HookUseWindowSizePage';
 
+export * from './ServiceClipboardPage';
+export * from './ServiceFilePage';
+export * from './ServiceLocalPage';
+export * from './ServiceLoggerPage';
+
 export * from '../assets';